improved map collisions

+++ GameObject.h
@@ -4,11 +4,12 @@
 #include "TextureW.h"
 #include "Collider.h"
 #include <vector>
+#include <cmath>
 
 class GameObject
 {
 protected:
-    Vector2D<double> m_position;
+    Vector2D<float> m_position;
 
     TextureW m_texture;
     SDL_Rect m_srcRect;
@@ -17,7 +18,7 @@ protected:
     Collider m_collider;
 
 public:
-    GameObject(const char* fileName, double xPos = 0, double yPos = 0, int colliderWidth = 0, int colliderHeight = 0);
+    GameObject(const char* fileName, float xPos = 0, float yPos = 0, int colliderWidth = 0, int colliderHeight = 0);
     ~GameObject();
 
     virtual void draw();
