+++ TextureW.h
@@ -11,7 +11,7 @@ public:
     ~TextureW();
 
     void load(const char* fileName);
-    void draw(SDL_Rect srcRect, SDL_Rect dstRect, double angle, SDL_Point* centre, SDL_RendererFlip flip);
+    void draw(SDL_Rect srcRect, SDL_Rect dstRect, float angle, SDL_Point* centre, SDL_RendererFlip flip);
     void free();
     void setColour(Uint8 R, Uint8 G, Uint8 B);
     void setBlendingMode(SDL_BlendMode mode);
